Avoid further implicit type conversions
Summary:
Some overflow was detected with D73022 and this diff fixes the occurences in PopLibs

Ref T67303

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, Eddy

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, Eddy

Subscribers: Eddy

Maniphest Tasks: T67303

Differential Revision: https://phabricator.sourcevertex.net/D73399

+++ lib/popsparse/HyperGraph.cpp
@@ -41,7 +41,7 @@ void HyperGraph::addConv1x1Vertex(poplar::Graph &graph,
 
     int workerSize = batchSize / nWorker;
     int leftover = batchSize % nWorker;
-    std::vector<int> worklist(nWorker * 3);
+    std::vector<unsigned short> worklist(nWorker * 3);
     int offset = 0;
 
     // Distrubute batchSize between workers as even as possible
